% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_template.r
\name{get_max_PurchAmount}
\alias{get_max_PurchAmount}
\title{get_max_PurchAmount}
\usage{
get_max_PurchAmount(x, date)
}
\arguments{
\item{x}{a data object with columns TransDate, Customer, PurchAmount, Cost}

\item{date}{a specification for the time of interest. Must be of format "dd.mm.yyy"}
}
\value{
The Customer with the highest profit for a given date \code{date} of data \code{x}
}
\description{
Function that returns the highest PurchAmount for a given date.
}
\details{
\code{data} contains the transaction data. The data set must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
\examples{
data(transactions)
myFun(transactions, date="01.10.2020")

}
