% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.r
\name{RFMfunction}
\alias{RFMfunction}
\title{RFMfunction}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{arg1}{data}

\item{arg2}{a number}

\item{arg3}{a number}

\item{arg4}{a number}
}
\value{
the final result of \code{arg1},\code{arg2},\code{arg3},\code{arg4}
}
\description{
This function calculates the RFM score
}
\details{
\code{data} contains the transaction data. Must contain columns 'Customer', 'TransData', and 'PurchAmount'
}
\examples{
RFMfunction(data, 20, 20, 60)

}
